# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: ''
  title: Jongdeug's Blog API Docs
  description: 블로그 API 명세서입니다.
  license:
    name: Jongdeug
servers:
  - url: 'http://localhost:8000/'
paths:
...
  /posts:
    post:
      tags:
        - Posts
      summary: 게시글 생성
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreatePostDto'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    description: 게시글 id
                    example: '12345'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenAPIError'
        '404':
          $ref: '#/components/responses/NotFoundUserError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/posts/{id}':
    patch:
      tags:
        - Posts
      summary: 게시글 수정
      parameters:
        - name: id
          in: path
          required: true
          description: 수정할 게시글 id
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdatePostDto'
      responses:
        '200':
          $ref: '#/components/responses/SuccessOkResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 해당 API 를 사용할 권한이 없습니다 or 게시글에 대한 권한이 없습니다
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 유저를 찾을 수 없습니다 or 게시글을 찾을 수 없습니다
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    CreatePostDto:
      type: object
      properties:
        title:
          type: string
          description: 제목
          example: 생성된 게시글
        content:
          type: string
          description: 내용
          example: 생성했습니다
        category:
          type: string
          description: 카테고리
          example: 운영체제
        tags:
          type: array
          items:
            type: string
          description: 태그
          example:
            - 커널
            - 스케줄링
        images:
          type: array
          items:
            type: string
            format: binary
          description: 이미지 파일
      required:
        - title
        - content
        - category
    UpdatePostDto:
      type: object
      properties:
        title:
          type: string
          description: 제목
          example: 수정된 게시글
        content:
          type: string
          description: 내용
          example: 수정했습니다
        category:
          type: string
          description: 카테고리
          example: 알고리즘
        tags:
          type: array
          items:
            type: string
          description: 태그
          example:
            - 리눅스
            - 시간복잡도
        images:
          type: array
          items:
            type: string
            format: binary
          description: 이미지 파일
      required:
        - title
        - content
        - category
  responses:
...
    SuccessOkResponse:
      description: OK
      content:
        application/json:
          schema:
            type: object
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 서버에서 에러가 발생했습니다
    ForbiddenAPIError:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 해당 API 를 사용할 권한이 없습니다
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 잘못된 토큰입니다 or 만료된 토큰입니다 or 토큰을 보내고 있지 않습니다 or 서비스 이용은 access 토큰으로만 가능합니다 or 로그인을 진행해주세요
...
    NotFoundUserError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 유저를 찾을 수 없습니다